Refactor Solutions, SolutionListPage and CaseHistories codecs
