+++ .changeset/sour-rocks-greet.md
@@ -1,0 +1,5 @@
+---
+"nextjs-website": patch
+---
+
+Refactor Solutions, SolutionListPage and CaseHistories codecs
