+++ apps/nextjs-website/src/lib/cmsApi.ts
@@ -23,11 +23,11 @@ import {
   tutorialLists,
 } from '@/_contents/products';
 import { makeCaseHistoriesProps } from './caseHistories';
-import { fetchCaseHistories } from './strapi/caseHistoriesCodec';
-import { fetchSolutions } from './strapi/solutionsCodec';
+import { fetchCaseHistories } from './strapi/fetches/fetchCaseHistories';
+import { fetchSolutions } from './strapi/fetches/fetchSolutions';
 import { makeDetailSolutionsProps, makeFullSolutionsProps } from './solutions';
 import { makeSolutionListProps } from './solutionList';
-import { fetchSolutionList } from './strapi/solutionListCodec';
+import { fetchSolutionListPage } from './strapi/fetches/fetchSolutionListPage';
 import { fetchApiDataListPages } from './strapi/fetches/fetchApiDataListPages';
 import { makeApiDataListPageProps } from './apiDataListPages';
 import { makeApiDataProps } from './apiDataPages';
@@ -199,7 +199,7 @@ export const getSolutionsListProps = async () => {
   } = buildEnv;
 
   if (fetchFromStrapi) {
-    const strapiSolutionsList = await fetchSolutionList(buildEnv);
+    const strapiSolutionsList = await fetchSolutionListPage(buildEnv);
     return makeSolutionListProps(strapiSolutionsList);
   }
 };
