+++ apps/nextjs-website/src/lib/strapi/codecs/CaseHistoriesCodec.ts
@@ -1,0 +1,32 @@
+import * as t from 'io-ts/lib';
+import * as tt from 'io-ts-types';
+import { NullToUndefinedCodec } from './NullToUndefinedCodec';
+import { MediaCodec } from './MediaCodec';
+import { PaginationCodec } from './PaginationCodec';
+import { ProductCodec } from './ProductCodec';
+import { PartCodec } from './PartCodec';
+
+export const CaseHistoryCodec = t.strict({
+  id: t.number,
+  attributes: t.strict({
+    slug: t.string,
+    title: t.string,
+    description: t.union([NullToUndefinedCodec, t.string]),
+    publishedAt: tt.DateFromISOString,
+    image: t.union([
+      NullToUndefinedCodec,
+      t.strict({ data: t.union([NullToUndefinedCodec, MediaCodec]) }),
+    ]),
+    products: t.strict({
+      data: t.array(ProductCodec),
+    }),
+    parts: t.array(PartCodec),
+  }),
+});
+
+export const CaseHistoriesCodec = t.strict({
+  data: t.array(CaseHistoryCodec),
+  meta: PaginationCodec,
+});
+
+export type StrapiCaseHistories = t.TypeOf<typeof CaseHistoriesCodec>;
